requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
greenlet>=3.0.0
psycopg2-binary==2.9.10
pgvector==0.3.2
alembic==1.13.3

# AWS SDK
boto3==1.35.36

# Git operations
gitpython==3.1.43

# Markdown parsing
markdown==3.7
beautifulsoup4==4.12.3

# Word document parsing
python-docx==1.1.2

# Config & validation
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Reranking (CPU only)
sentence-transformers>=3.0.0
--extra-index-url https://download.pytorch.org/whl/cpu

# Resilience
tenacity==9.0.0

# CLI
click==8.1.7

# Utilities
python-multipart==0.0.12
httpx==0.27.2
